Lablib Data Files
Detect 32- or 64-bit mode and change the file format specifier to indicate which.

Changes to DataConvert to detect when it had a file in the wrong mode.

+++ Applications/DataConvert/PrintView.m
@@ -95,7 +95,7 @@ static long			printWidthPix;
     return YES;
 }
 
-- (NSRect)rectForPage:(long)pageNum
+- (NSRect)rectForPage:(NSInteger)pageNum
 {
     NSRect theRect;
